Change how mods are added to the server
Mods will be added to the log-watcher for copying to the correct Minecraft server path. This may lead to requiring a restart of the Minecraft server, but it will actually work correctly every time now after restarting. Before, mods would _sometimes_ get added depending on how the volume felt about adding files to it